style: implement upload document modal scrollbar and move the UploadDocumentModal component
