+++ components/HomePage/Kanban/Column/JobPosts/JobModal/JobDocuments/Documents.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'next/navigation';
 
 import documents from '@/data/documents';
-import UploadDocumentModal from '@/components/Misc/UploadDocumentModal';
+import UploadDocumentModal from '@/components/Forms/AddDocument/UploadDocumentModal';
 import { LinkDocument } from '@/components/HomePage/HomeNavbar/LinkDocument';
 
 const Documents = () => {
